+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Skip updating binaries built from source.
+
+  Binaries built from source (either using `go build main.go` or `go install`
+  in the cloned repository) likely have been modified prior to being built.
+  Updating them would likely throw away these changes and end up being annoying
+  for engineers who want to keep their modified versions.
+
+  Moreover, packages built using `go build main.go` have
+  `command-line-arguments` set as their `path` in `go version -m binary-name`.
+  This makes it impossible to update automatically.
+
+- Filesystem path handling on Windows.
+
+  Use correct separator for filesystem paths on Windows. This allows using this
+  tool on Windows.
+
 ## v0.1.0 (2022-03-14)
 
 - Complete the basic functionality of upgrading globally installed executables.